1.0.9 1.增加每週運動提醒的功能。 2.修正進入會員資料編輯的頁面以及iPAD下點擊分享按鈕會crash的BUG

+++ TrainingRecord/Model/MemberUserDataToSystem.swift
@@ -22,8 +22,8 @@ class MemberUserDataToFirestore {
         
         if (memberDatas["userAge"] as! String) != ""  {
             let nowTime = Date().timeIntervalSinceNow
-            let recordTime = (memberDatas["userRecordTime"]as! [String]).first
-            let timeDifference = Int(nowTime) - Int(recordTime!)!
+            let recordTime = (memberDatas["userRecordTime"] as! [String]).first
+            let timeDifference = nowTime - Double(recordTime!)!
             
             var gapYearOld = timeDifference/(60*60*24*365)
             var yearDifference = 0
